+++ backend/src/main/java/com/bupt/backend/service/PostService.java
@@ -9,10 +9,10 @@ import java.util.List;
 public interface PostService {
     Result<Post> createPost(Post post, MultipartFile coverImage);
     Result<Post> updatePost(Post post, MultipartFile coverImage);
-    Result<Post> publishPost(Integer postId);
-    Result<Post> getPostById(Integer postId);
-    Result<List<Post>> getRecentPosts(Integer userId, Integer limit);
-    Result<List<Post>> getPostsByCategory(Integer categoryId);
-    Result<String> uploadCoverImage(Integer userId, MultipartFile file);
-    Result<Void> incrementViewCount(Integer postId);
+    Result<Post> publishPost(Long postId);
+    Result<Post> getPostById(Long postId);
+    Result<List<Post>> getRecentPosts(Long userId, Integer limit);
+    Result<List<Post>> getPostsByCategory(Long categoryId);
+    Result<String> uploadCoverImage(Long userId, MultipartFile file);
+    Result<Void> incrementViewCount(Long postId);
 }
